+++ src/features/user/validations/index.ts
@@ -7,6 +7,7 @@ export const createUserSchema = z.object({
   lastName: z.string().min(1),
   password: z.string().min(8),
   role: z.enum(['user', 'company']),
+  autoConfirmEmail: z.boolean().optional(),
 });
 
 export const getUserByIdSchema = z.object({
